:truck: Move utility function to correct location
* everything notifications related should be contained in the
  notifications app
* leaking optional apps into the core app breaks projects that don't
  have the notifications app enabled
