+++ vng_api_common/utils.py
@@ -12,8 +12,6 @@ from django.utils.module_loading import import_string
 
 from zds_client.client import ClientError
 
-from vng_api_common.notifications.kanalen import Kanaal
-
 try:
     from djangorestframework_camel_case.util import underscore_to_camel as _underscore_to_camel
 except ImportError:
@@ -150,18 +148,3 @@ def get_help_text(model_string: str, field_name: str) -> str:
     ModelClass = apps.get_model(model_string, require_ready=False)
     field = ModelClass._meta.get_field(field_name)
     return field.help_text
-
-
-def notification_documentation(kanaal: Kanaal):
-    """
-    Generate notification documentation for an OpenAPI specification containing
-    the relevant resources and actions for a given KANAAL
-    """
-    doc = f"""### Notificaties \
-           \n\nDeze API publiceert notificaties op het kanaal `{kanaal.label}`. \
-           \n\n{kanaal.description}
-           \n\n**Resources en acties**\n\n"""
-
-    for resource, actions in kanaal.get_usage():
-        doc += f"""- `{resource}`: {', '.join(actions)}\n"""
-    return doc
